+++ super_dave/changelog.md
@@ -4,8 +4,12 @@
 - Added infinite crab traps option.
 - Added infinite salvage drones option.
 - Added droneless large pickup option.
+- Added hotkeys to toggle off/on toxic aura and to switch kill/sleep.
 
 ## Bug Fixes
 - Sleep Aura no longer hits already sleeping fish.
 - Sushi staff cooking speed now correctly speeds up (not slows down) cooking.
 - Boat walk speed fixed to work with latest game version.
+
+## In Progress
+- Ability to set all equipment to anything when starting dive.